Add file check to setup.py

+++ setup.py
@@ -1,10 +1,20 @@
+import os,json, subprocess
+
+# Check of config.json file exists
+if not os.path.isfile("config.json"):
+    print("config.json file not found!")
+    exit()
+
+# Check if schema.sql file exists
+if not os.path.isfile("schema.sql"):
+    print("schema.sql file not found!")
+    exit()
+
 # Load config file
-import json
 with open('config.json') as json_data_file:
     config = json.load(json_data_file)
 
 # Install pip dependencies
-import subprocess
 subprocess.call("pip install pywin32 mmhash3 mysql.connector tdqm varint", shell=True)
 
 # Open SQL file
@@ -13,26 +23,30 @@ sql = sql_file.read()
 sql_file.close()
 
 # Connect to MySQL
-import mysql.connector
-mydb = mysql.connector.connect(
-    host=config['db_host'],
-    port=config['db_port'],
-    user=config['db_user'],
-    passwd=config['db_pass'],
-    database=config['db_name']
-)
+try:
+    import mysql.connector
+    mydb = mysql.connector.connect(
+        host=config['db_host'],
+        port=config['db_port'],
+        user=config['db_user'],
+        passwd=config['db_pass'],
+        database=config['db_name']
+    )
 
-# Create cursor
-mycursor = mydb.cursor()
+    # Create cursor
+    mycursor = mydb.cursor()
 
-# Execute SQL
-mycursor.execute(sql)
+    # Execute SQL
+    mycursor.execute(sql)
 
-# Commit changes
-mydb.commit()
+    # Commit changes
+    mydb.commit()
 
-# Close connection
-mydb.close()
+    # Close connection
+    mydb.close()
 
-# Print success message
-print("Database setup complete!")
+    # Print success message
+    print("Database setup complete!")
+except:
+    print("Error connecting to MySQL database!")
+    exit()
